+++ nahye/mission12/Main.java
@@ -1,15 +1,16 @@
 package mission12;
 
+import mission12.controller.MenuController;
 import mission12.service.LibraryService;
 import mission12.service.UserService;
 import mission12.view.MenuView;
 
 public class Main {
-
     public static void main(String[] args) {
-        LibraryService libraryService = new LibraryService();
-        UserService userService = new UserService();
-        MenuView view = new MenuView(libraryService, userService);
+        MenuController menuController = new MenuController(
+                new LibraryService(), new UserService());
+
+        MenuView view = new MenuView(menuController);
         view.showMainView();
     }
 }
