+++ nahye/mission12/user/User.java
@@ -3,19 +3,15 @@ package mission12.user;
 import mission12.view.MenuView;
 
 public interface User {
-    public String name = "";
-    public String id = "";
-    public String password = "";
+
+    String getId();
+
+    String getName();
+
+    boolean hasUserById(String id);
 
 
-    public String getId();
-
-    public String getName();
-
-    public boolean hasUserById(String id);
-
-
-    public boolean checkPassword(String password);
+    boolean checkPassword(String password);
 
     void showMenuView(MenuView view);
